requirements.txt
asgiref==3.8.1
attrs==25.3.0
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==2.0.12
cryptography==45.0.4
defusedxml==0.7.1
Django==5.2.3
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
djoser==2.3.1
idna==3.10
iniconfig==2.1.0
oauthlib==3.2.2
packaging==25.0
Pillow==9.3.0
pluggy==0.13.1
py==1.11.0
pycparser==2.22
PyJWT==2.9.0
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
python3-openid==3.2.0
pytz==2025.2
requests==2.26.0
requests-oauthlib==2.0.0
drf-nested-routers==0.93.5
social-auth-app-django==5.4.3
social-auth-core==4.6.1
sqlparse==0.5.3
toml==0.10.2
urllib3==1.26.20
